# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Publication API",
        "description": "API pour gérer les publications (texte, images et vidéos) les commentaires et les likes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Server local"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/publications/video": {
            "post": {
                "tags": [
                    "Publications"
                ],
                "summary": "Créer une publication de type vidéo",
                "operationId": "d0560891198d79d2bf7f38de892adcce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "video_file",
                                    "title"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID de l'utilisateur qui cree la publication",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "video_file": {
                                        "description": "Fichier a uploader",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Titre de la publication"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description de la publication"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Publication vidéo créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publication"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/publications/text-with-images": {
            "post": {
                "tags": [
                    "Publications"
                ],
                "summary": "Créer une publication de type texte avec images",
                "description": "Crée une publication contenant du texte et une ou plusieurs images. Les images doivent être au format JPEG, PNG, JPG ou GIF, avec une taille maximale de 2 Mo par image.",
                "operationId": "cbba4a9e3dfb5ceca1757810b40db5d3",
                "requestBody": {
                    "description": "Données nécessaires pour créer une publication texte + images",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "content",
                                    "images",
                                    "title"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID de l'utilisateur créant la publication",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "description": "Contenu principal de la publication",
                                        "type": "string",
                                        "example": "Contenu textuel de la publication"
                                    },
                                    "images": {
                                        "description": "Liste des images à associer à la publication",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "title": {
                                        "description": "Titre de la publication",
                                        "type": "string",
                                        "example": "Titre de la publication"
                                    },
                                    "description": {
                                        "description": "Description optionnelle de la publication",
                                        "type": "string",
                                        "example": "Description de la publication"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Publication texte + images créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Publication créée avec succès"
                                        },
                                        "publication": {
                                            "$ref": "#/components/schemas/Publication"
                                        },
                                        "images": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PublicationImage"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation ou autre erreur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Une erreur s'est produite lors de la création de la publication."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur serveur interne"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/publications/author/{authorId}": {
            "get": {
                "tags": [
                    "Publications"
                ],
                "summary": "Récupérer toutes les publications d'un auteur",
                "operationId": "4f86d10ff5e842130be2a146e7f9a81a",
                "parameters": [
                    {
                        "name": "authorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des publications de l'auteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Publication"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Auteur non trouvé"
                    }
                }
            }
        },
...
        "/publications": {
            "get": {
                "tags": [
                    "Publications"
                ],
                "summary": "Récupérer toutes les publications",
                "operationId": "5198364499a77a77f94be76b2ec7fc9e",
                "responses": {
                    "200": {
                        "description": "Liste des publications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Publications récupérées avec succès."
                                        },
                                        "publications": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Publication"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/publications/video/{filename}": {
            "get": {
                "tags": [
                    "Publications"
                ],
                "summary": "Streamer une vidéo",
                "operationId": "fa24eab81d4e34daa053fe11c76213ba",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Nom du fichier vidéo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vidéo streamée avec succès",
                        "content": {
                            "video/mp4": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vidéo non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PublicationImage": {
                "description": "Représente une image associée à une publication",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "publication_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_path": {
                        "type": "string",
                        "example": "public/publications/images/abc123.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}